[Go functions] access to user config before function loop (#8467)
The pulsar go functions SDK has no access to user config outside of the function context, which is only accessible from inside the function handler itself (i.e. once processing has started). This is a somewhat awkward access method when it comes to user configuration values which might be used to initialize a persistent connection to a database or other service, since it would require a largely-superfluous check for the initialization and related configuration values in every run of the function, when it's only required once at startup.

This PR is a re-hash of #8365, which introduced secrets access incompatible to the concept of a secretsprovider, which will need to be addressed as separate parity functionality.